* Consolidated styling: Pass 1
