+++ app/(tabs)/dragdrop.tsx
@@ -1,12 +1,4 @@
 import React from "react";
-import { StyleSheet } from "react-native";
-import { Menu, Divider, Button, MD3LightTheme } from "react-native-paper";
-import {
-  heightPercentageToDP,
-  widthPercentageToDP,
-} from "react-native-responsive-screen";
-
-import { View } from "@/components/Themed";
 import ScrapbookEntry from "@/components/Scrapbooking/ScrapbookEntry";
 
 
@@ -18,12 +10,3 @@ export default function DragDrop() {
    <ScrapbookEntry />
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    paddingTop: 0,
-    flex: 1,
-    justifyContent: "center",
-    backgroundColor: MD3LightTheme.colors.background,
-  },
-});
