+++ app/(tabs)/mentalhelp.tsx
@@ -1,19 +1,20 @@
 import React from 'react';
-import { View, StyleSheet } from 'react-native';
+import { View } from 'react-native';
 import { Card, Text  } from 'react-native-paper';
 import MentalHealthComponent from '@/components/MentalHealth/MentalHealth';
+import { MentalHelpStyles } from '@/styles/styles';
 
 const MentalHelp: React.FC = () => {
     return (
-        <View style={styles.container}>
-            <Card style={styles.cardTwo}>
+        <View style={MentalHelpStyles.container}>
+            <Card style={MentalHelpStyles.cardTwo}>
                 <Card.Content>
-                    <Text style={styles.description}>
+                    <Text style={MentalHelpStyles.description}>
                         Here are some resources to help you with your mental health.
                         </Text>
                 </Card.Content>
             </Card>
-            <Card style={styles.card}>
+            <Card style={MentalHelpStyles.card}>
                 <Card.Content>
                     <MentalHealthComponent />
                 </Card.Content>
@@ -21,24 +22,5 @@ const MentalHelp: React.FC = () => {
         </View>
     );
 }
-
-const styles = StyleSheet.create({
-    container: {
-        flex: 1,
-        padding: 20,
-        backgroundColor: '#f0f0f0',
-    },
-    card: {
-        marginBottom: 20,
-    },
-    cardTwo: {
-        marginBottom: 20,
-    },
-    description: {
-        fontSize: 16,
-        marginBottom: 10,
-        fontWeight: 'bold',
-    },
-});
 
 export default MentalHelp;
