+++ app/(tabs)/profile.tsx
@@ -1,7 +1,5 @@
-import InventoryCard from '@/components/Store/InventoryCard';
 import ProfileCard from '@/components/Profile/ProfileCard';
 import React from 'react';
-import { View, Text } from 'react-native';
 import { ScrollView } from 'react-native-gesture-handler';
 import { widthPercentageToDP } from 'react-native-responsive-screen';
 import {LinearGradient} from 'expo-linear-gradient';
@@ -11,9 +9,7 @@ const Profile = () => {
         <>
         <LinearGradient colors={['#4c669f', '#3b5998', '#192f6a']} style={{width:widthPercentageToDP(100), position:'absolute', top:0, left:0, right:0, bottom:0}}/>
         <ScrollView contentContainerStyle={{  alignItems: 'center', justifyContent: 'center', top:0 }}>
-            
            <ProfileCard/>
-           
         </ScrollView></>
     );
 };
