+++ components/EditScreenInfo.tsx
@@ -1,18 +1,17 @@
 import React from "react";
-import { StyleSheet } from "react-native";
-
 import { ExternalLink } from "./ExternalLink";
 import { MonoText } from "./StyledText";
 import { Text, View } from "./Themed";
+import { EditScreenInfoStyles } from "@/styles/styles";
 
 import Colors from "@/constants/Colors";
 
 export default function EditScreenInfo({ path }: { path: string }) {
   return (
     <View>
-      <View style={styles.getStartedContainer}>
+      <View style={EditScreenInfoStyles.getStartedContainer}>
         <Text
-          style={styles.getStartedText}
+          style={EditScreenInfoStyles.getStartedText}
           lightColor="rgba(0,0,0,0.8)"
           darkColor="rgba(255,255,255,0.8)"
         >
@@ -20,7 +19,7 @@ export default function EditScreenInfo({ path }: { path: string }) {
         </Text>
 
         <View
-          style={[styles.codeHighlightContainer, styles.homeScreenFilename]}
+          style={[EditScreenInfoStyles.codeHighlightContainer, EditScreenInfoStyles.homeScreenFilename]}
           darkColor="rgba(255,255,255,0.05)"
           lightColor="rgba(0,0,0,0.05)"
         >
@@ -28,7 +27,7 @@ export default function EditScreenInfo({ path }: { path: string }) {
         </View>
 
         <Text
-          style={styles.getStartedText}
+          style={EditScreenInfoStyles.getStartedText}
           lightColor="rgba(0,0,0,0.8)"
           darkColor="rgba(255,255,255,0.8)"
         >
@@ -37,12 +36,12 @@ export default function EditScreenInfo({ path }: { path: string }) {
         </Text>
       </View>
 
-      <View style={styles.helpContainer}>
+      <View style={EditScreenInfoStyles.helpContainer}>
         <ExternalLink
-          style={styles.helpLink}
+          style={EditScreenInfoStyles.helpLink}
           href="https://docs.expo.io/get-started/create-a-new-app/#opening-the-app-on-your-phonetablet"
         >
-          <Text style={styles.helpLinkText} lightColor={Colors.light.tint}>
+          <Text style={EditScreenInfoStyles.helpLinkText} lightColor={Colors.light.tint}>
             Tap here if your app doesn't automatically update after making
             changes
           </Text>
@@ -51,33 +50,3 @@ export default function EditScreenInfo({ path }: { path: string }) {
     </View>
   );
 }
-
-const styles = StyleSheet.create({
-  getStartedContainer: {
-    alignItems: "center",
-    marginHorizontal: 50,
-  },
-  homeScreenFilename: {
-    marginVertical: 7,
-  },
-  codeHighlightContainer: {
-    borderRadius: 3,
-    paddingHorizontal: 4,
-  },
-  getStartedText: {
-    fontSize: 17,
-    lineHeight: 24,
-    textAlign: "center",
-  },
-  helpContainer: {
-    marginTop: 15,
-    marginHorizontal: 20,
-    alignItems: "center",
-  },
-  helpLink: {
-    paddingVertical: 15,
-  },
-  helpLinkText: {
-    textAlign: "center",
-  },
-});
